% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senateaR-package.R
\docType{package}
\name{senateaR-package}
\alias{senateaR}
\alias{senateaR-package}
\title{senateaR: Access Open Data from the Argentine Senate}
\description{
Provides functions to fetch and parse open data published by the Argentine Senate ("Senado de la Nacion Argentina"), including current and historical senators, voting records and vote details, commissions, session bulletins, procurement notices ("licitaciones"), parliamentary entries ("asuntos entrados"), and stenographic transcripts. Data are retrieved directly from the Senate's public open-data portal and website.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vsntos/senateaR}
  \item Report bugs at \url{https://github.com/vsntos/senateaR/issues}
}

}
\author{
\strong{Maintainer}: Vinicius Santos \email{santos.vinicius18@gmail.com}

}
\keyword{internal}
